backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
python-dotenv
requests
psycopg2-binary
fpdf2
google-generativeai
gTTS